compile error fix

+++ chp3/6_bottom_up_migration_before/calculator/src/com/packt/calculator/commands/SimpleInterestCommand.java
@@ -7,7 +7,7 @@ public class SimpleInterestCommand implements Command{
 	public final Integer rate; 
 	public final Integer duration;
 
-	public SimpleInterestCommand(Integer principal, Integer rate, Integer duration){
+	public SimpleInterestCommand(Double principal, Integer rate, Integer duration){
 		this.principal = principal;
 		this.rate = rate;
 		this.duration = duration;
